templates/longevity/requirements.txt
algorithmia>=1.0.0,<2.0
six
scipy==1.4.1
matplotlib==3.2.1
tqdm==4.46.0
numpy==1.18.4
opencv-python==4.2.0.34
pillow==7.1.2
tensorboard==2.2.1
torch>=1.7.0,<1.8.0
torchvision==0.4.2

# asli - additions
cmake
dlib
ninja
